@@ -1,6 +1,6 @@
 apply plugin: 'com.android.application'
 def AAVersion = '4.6.0'
-def AndroidSupportVersion = '26.1.0'
+def AndroidSupportVersion = '1.0.0-beta01'
 
 buildscript {
     repositories {
@@ -9,7 +9,7 @@ buildscript {
 }
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 28
 
     defaultConfig {
         applicationId "com.hiddenramblings.tagmo"
@@ -61,16 +61,16 @@ android {
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    implementation "com.android.support:design:$AndroidSupportVersion"
-    implementation "com.android.support:appcompat-v7:$AndroidSupportVersion"
-    implementation "com.android.support:recyclerview-v7:$AndroidSupportVersion"
-    implementation "com.android.support:preference-v7:$AndroidSupportVersion"
-    implementation "com.android.support:preference-v14:$AndroidSupportVersion"
-    implementation "com.android.support:cardview-v7:$AndroidSupportVersion"
+    implementation "com.google.android.material:material:$AndroidSupportVersion"
+    implementation "androidx.appcompat:appcompat:$AndroidSupportVersion"
+    implementation "androidx.recyclerview:recyclerview:$AndroidSupportVersion"
+    implementation "androidx.preference:preference:$AndroidSupportVersion"
+    implementation "androidx.legacy:legacy-preference-v14:$AndroidSupportVersion"
+    implementation "androidx.cardview:cardview:$AndroidSupportVersion"
     annotationProcessor "org.androidannotations:androidannotations:$AAVersion"
     implementation "org.androidannotations:androidannotations-api:$AAVersion"
-    annotationProcessor 'org.parceler:parceler:1.1.9'
-    implementation 'org.parceler:parceler-api:1.1.9'
+    annotationProcessor 'org.parceler:parceler:1.1.12'
+    implementation 'org.parceler:parceler-api:1.1.12'
     implementation 'com.robertlevonyan.view:MaterialChipView:1.0.1'
     implementation 'org.apmem.tools:layouts:1.10@aar'
     implementation 'com.github.bumptech.glide:glide:4.0.0'
