@@ -1,6 +1,6 @@
 package com.hiddenramblings.tagmo;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 public abstract class AppDataFragment extends Fragment implements AppDataInterface {
 }
