@@ -1,7 +1,7 @@
 package com.hiddenramblings.tagmo;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.widget.ArrayAdapter;
 import android.widget.EditText;
 import android.widget.Spinner;
