@@ -2,8 +2,8 @@ package com.hiddenramblings.tagmo;
 
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 
 import org.androidannotations.annotations.EActivity;
 
